Refactoring useGasFeeInputs hook
